@@ -3,10 +3,7 @@ Python-Redmine tries it's best to provide human readable errors in all situation
 This is a list of all exceptions or warnings that Python-Redmine can throw/raise.
 """
 
-from . import utilities
 
-
-@utilities.fix_unicode
 class BaseRedmineWarning(Warning):
     """
     Base warning class for Redmine warnings.
@@ -19,7 +16,6 @@ class PerformanceWarning(BaseRedmineWarning):
     """
 
 
-@utilities.fix_unicode
 class BaseRedmineError(Exception):
     """
     Base exception class for Redmine exceptions.
@@ -31,7 +27,7 @@ class ResourceError(BaseRedmineError):
     Unsupported Redmine resource exception.
     """
     def __init__(self):
-        super(ResourceError, self).__init__('Unsupported Redmine resource')
+        super().__init__('Unsupported Redmine resource')
 
 
 class NoFileError(BaseRedmineError):
@@ -39,7 +35,7 @@ class NoFileError(BaseRedmineError):
     File doesn't exist or is empty exception.
     """
     def __init__(self):
-        super(NoFileError, self).__init__("Can't upload a file that doesn't exist or is empty")
+        super().__init__("Can't upload a file that doesn't exist or is empty")
 
 
 class FileObjectError(BaseRedmineError):
@@ -47,7 +43,7 @@ class FileObjectError(BaseRedmineError):
     File-like object isn't supported as it doesn't support the read(size) method.
     """
     def __init__(self):
-        super(FileObjectError, self).__init__("File-like object doesn't support the read(size) method")
+        super().__init__("File-like object doesn't support the read(size) method")
 
 
 class ResourceNotFoundError(BaseRedmineError):
@@ -55,7 +51,7 @@ class ResourceNotFoundError(BaseRedmineError):
     Requested resource doesn't exist.
     """
     def __init__(self):
-        super(ResourceNotFoundError, self).__init__("Requested resource doesn't exist")
+        super().__init__("Requested resource doesn't exist")
 
 
 class ConflictError(BaseRedmineError):
@@ -63,7 +59,7 @@ class ConflictError(BaseRedmineError):
     Resource version on the server is newer than on the client.
     """
     def __init__(self):
-        super(ConflictError, self).__init__('Resource version on the server is newer than on the client')
+        super().__init__('Resource version on the server is newer than on the client')
 
 
 class AuthError(BaseRedmineError):
@@ -71,7 +67,7 @@ class AuthError(BaseRedmineError):
     Invalid authentication details.
     """
     def __init__(self):
-        super(AuthError, self).__init__('Invalid authentication details')
+        super().__init__('Invalid authentication details')
 
 
 class ImpersonateError(BaseRedmineError):
@@ -79,7 +75,7 @@ class ImpersonateError(BaseRedmineError):
     Invalid impersonate login provided.
     """
     def __init__(self):
-        super(ImpersonateError, self).__init__("Impersonate login provided doesn't exist or isn't active")
+        super().__init__("Impersonate login provided doesn't exist or isn't active")
 
 
 class ServerError(BaseRedmineError):
@@ -87,7 +83,7 @@ class ServerError(BaseRedmineError):
     Redmine internal error.
     """
     def __init__(self):
-        super(ServerError, self).__init__('Redmine returned internal error, check Redmine logs for details')
+        super().__init__('Redmine returned internal error, check Redmine logs for details')
 
 
 class RequestEntityTooLargeError(BaseRedmineError):
@@ -95,7 +91,7 @@ class RequestEntityTooLargeError(BaseRedmineError):
     Size of the request exceeds the capacity limit on the server.
     """
     def __init__(self):
-        super(RequestEntityTooLargeError, self).__init__(
+        super().__init__(
             "The requested resource doesn't allow POST requests or the size of the request exceeds the capacity limit")
 
 
@@ -105,8 +101,7 @@ class UnknownError(BaseRedmineError):
     """
     def __init__(self, status_code):
         self.status_code = status_code
-        super(UnknownError, self).__init__(
-            'Redmine returned unknown error with the status code {0}'.format(status_code))
+        super().__init__(f'Redmine returned unknown error with the status code {status_code}')
 
 
 class ValidationError(BaseRedmineError):
@@ -114,7 +109,7 @@ class ValidationError(BaseRedmineError):
     Redmine validation errors occurred on create/update resource.
     """
     def __init__(self, error):
-        super(ValidationError, self).__init__(error)
+        super().__init__(error)
 
 
 class ResourceSetIndexError(BaseRedmineError):
@@ -122,7 +117,7 @@ class ResourceSetIndexError(BaseRedmineError):
     Index doesn't exist in the ResourceSet.
     """
     def __init__(self):
-        super(ResourceSetIndexError, self).__init__('Resource not available by requested index')
+        super().__init__('Resource not available by requested index')
 
 
 class ResourceSetFilterLookupError(BaseRedmineError):
@@ -130,8 +125,7 @@ class ResourceSetFilterLookupError(BaseRedmineError):
     Resource set filter method received an invalid lookup in one of the filters.
     """
     def __init__(self, lookup, f):
-        super(ResourceSetFilterLookupError, self).__init__(
-            'Received an invalid lookup "{0}" in "{1}" filter'.format(lookup, f))
+        super().__init__(f'Received an invalid lookup "{lookup}" in "{f}" filter')
 
 
 class ResourceBadMethodError(BaseRedmineError):
@@ -139,7 +133,7 @@ class ResourceBadMethodError(BaseRedmineError):
     Resource doesn't support the requested method.
     """
     def __init__(self):
-        super(ResourceBadMethodError, self).__init__("Resource doesn't support the requested method")
+        super().__init__("Resource doesn't support the requested method")
 
 
 class ResourceFilterError(BaseRedmineError):
@@ -147,7 +141,7 @@ class ResourceFilterError(BaseRedmineError):
     Resource doesn't support requested filter(s).
     """
     def __init__(self):
-        super(ResourceFilterError, self).__init__("Resource doesn't support requested filter(s)")
+        super().__init__("Resource doesn't support requested filter(s)")
 
 
 class ResourceNoFiltersProvidedError(BaseRedmineError):
@@ -155,7 +149,7 @@ class ResourceNoFiltersProvidedError(BaseRedmineError):
     No filter(s) provided.
     """
     def __init__(self):
-        super(ResourceNoFiltersProvidedError, self).__init__('Resource needs some filters to be filtered on')
+        super().__init__('Resource needs some filters to be filtered on')
 
 
 class ResourceNoFieldsProvidedError(BaseRedmineError):
@@ -163,8 +157,7 @@ class ResourceNoFieldsProvidedError(BaseRedmineError):
     No field(s) provided.
     """
     def __init__(self):
-        super(ResourceNoFieldsProvidedError, self).__init__(
-            'Resource needs some fields to be set to be created/updated')
+        super().__init__('Resource needs some fields to be set to be created/updated')
 
 
 class ResourceAttrError(BaseRedmineError, AttributeError):
@@ -172,7 +165,7 @@ class ResourceAttrError(BaseRedmineError, AttributeError):
     Resource doesn't have the requested attribute.
     """
     def __init__(self):
-        super(ResourceAttrError, self).__init__("Resource doesn't have the requested attribute")
+        super().__init__("Resource doesn't have the requested attribute")
 
 
 class ReadonlyAttrError(BaseRedmineError):
@@ -180,7 +173,7 @@ class ReadonlyAttrError(BaseRedmineError):
     Resource can't set attribute that is read only.
     """
     def __init__(self):
-        super(ReadonlyAttrError, self).__init__("Can't set read only attribute")
+        super().__init__("Can't set read only attribute")
 
 
 class VersionMismatchError(BaseRedmineError):
@@ -188,7 +181,7 @@ class VersionMismatchError(BaseRedmineError):
     Feature isn't supported on specified Redmine version.
     """
     def __init__(self, feature):
-        super(VersionMismatchError, self).__init__("{0} isn't supported on specified Redmine version".format(feature))
+        super().__init__(f"{feature} isn't supported on specified Redmine version")
 
 
 class ResourceVersionMismatchError(VersionMismatchError):
@@ -196,7 +189,7 @@ class ResourceVersionMismatchError(VersionMismatchError):
     Resource isn't supported on specified Redmine version.
     """
     def __init__(self):
-        super(ResourceVersionMismatchError, self).__init__('Resource')
+        super().__init__('Resource')
 
 
 class ResultSetTotalCountError(BaseRedmineError):
@@ -204,7 +197,7 @@ class ResultSetTotalCountError(BaseRedmineError):
     ResultSet hasn't been yet evaluated and cannot yield a total_count.
     """
     def __init__(self):
-        super(ResultSetTotalCountError, self).__init__('Total count is unknown before evaluation')
+        super().__init__('Total count is unknown before evaluation')
 
 
 class CustomFieldValueError(BaseRedmineError):
@@ -212,7 +205,7 @@ class CustomFieldValueError(BaseRedmineError):
     Custom fields should be passed as a list of dictionaries.
     """
     def __init__(self):
-        super(CustomFieldValueError, self).__init__(
+        super().__init__(
             "Custom fields should be passed as a list of dictionaries in the form of [{'id': 1, 'value': 'foo'}]")
 
 
@@ -221,8 +214,7 @@ class ResourceRequirementsError(BaseRedmineError):
     Resource requires specified Redmine plugin(s) to function.
     """
     def __init__(self, requirements):
-        super(ResourceRequirementsError, self).__init__(
-            'The following requirements must be installed for resource to function: {0}'.format(
+        super().__init__('The following requirements must be installed for resource to function: {}'.format(
                 ', '.join(' >= '.join(req) if isinstance(req, (list, tuple)) else req for req in requirements)))
 
 
@@ -231,7 +223,7 @@ class FileUrlError(BaseRedmineError):
     URL provided to download a file can't be parsed.
     """
     def __init__(self):
-        super(FileUrlError, self).__init__("URL provided to download a file can't be parsed")
+        super().__init__("URL provided to download a file can't be parsed")
 
 
 class ForbiddenError(BaseRedmineError):
@@ -239,7 +231,7 @@ class ForbiddenError(BaseRedmineError):
     Requested resource is forbidden.
     """
     def __init__(self):
-        super(ForbiddenError, self).__init__('Requested resource is forbidden')
+        super().__init__('Requested resource is forbidden')
 
 
 class JSONDecodeError(BaseRedmineError):
@@ -248,7 +240,7 @@ class JSONDecodeError(BaseRedmineError):
     """
     def __init__(self, response):
         self.response = response
-        super(JSONDecodeError, self).__init__(
+        super().__init__(
             'Unable to decode received JSON, you can inspect exception\'s '
             '"response" attribute to find out what the response was')
 
@@ -258,7 +250,7 @@ class ExportNotSupported(BaseRedmineError):
     Export functionality not supported by resource.
     """
     def __init__(self):
-        super(ExportNotSupported, self).__init__('Export functionality not supported by resource')
+        super().__init__('Export functionality not supported by resource')
 
 
 class ExportFormatNotSupportedError(BaseRedmineError):
@@ -266,8 +258,7 @@ class ExportFormatNotSupportedError(BaseRedmineError):
     The given format isn't supported by resource.
     """
     def __init__(self):
-        super(ExportFormatNotSupportedError, self).__init__(
-            "The given format isn't supported by resource")
+        super().__init__("The given format isn't supported by resource")
 
 
 class HTTPProtocolError(BaseRedmineError):
@@ -275,7 +266,7 @@ class HTTPProtocolError(BaseRedmineError):
     Wrong HTTP protocol usage.
     """
     def __init__(self):
-        super(HTTPProtocolError, self).__init__('Redmine url should start with HTTPS and not with HTTP')
+        super().__init__('Redmine url should start with HTTPS and not with HTTP')
 
 
 class EngineClassError(BaseRedmineError):
@@ -283,4 +274,4 @@ class EngineClassError(BaseRedmineError):
     Engine isn't a class or isn't a BaseEngine subclass.
     """
     def __init__(self):
-        super(EngineClassError, self).__init__("Engine isn't a class or isn't a BaseEngine subclass")
+        super().__init__("Engine isn't a class or isn't a BaseEngine subclass")
