@@ -11,7 +11,7 @@ class FileManager(ResourceManager):
         if response is True:
             response = {self.container: {'id': int(request[self.container]['token'].split('.')[0])}}
 
-        return super(FileManager, self)._process_create_response(request, response)
+        return super()._process_create_response(request, response)
 
 
 class WikiPageManager(ResourceManager):
@@ -19,7 +19,7 @@ class WikiPageManager(ResourceManager):
         if response is True:
             raise exceptions.ValidationError('Resource already exists')  # issue #182
 
-        return super(WikiPageManager, self)._process_create_response(request, response)
+        return super()._process_create_response(request, response)
 
 
 class UserManager(ResourceManager):
@@ -31,18 +31,18 @@ class UserManager(ResourceManager):
         return path
 
     def _construct_get_url(self, path):
-        return super(UserManager, self)._construct_get_url(self._check_custom_url(path))
+        return super()._construct_get_url(self._check_custom_url(path))
 
     def _prepare_create_request(self, request):
-        request = super(UserManager, self)._prepare_create_request(request)
+        request = super()._prepare_create_request(request)
         request['send_information'] = request[self.container].pop('send_information', False)
         return request
 
     def _construct_update_url(self, path):
-        return super(UserManager, self)._construct_update_url(self._check_custom_url(path))
+        return super()._construct_update_url(self._check_custom_url(path))
 
     def _prepare_update_request(self, request):
-        request = super(UserManager, self)._prepare_update_request(request)
+        request = super()._prepare_update_request(request)
         request['send_information'] = request[self.resource_class.container_update].pop('send_information', False)
         return request
 
@@ -52,4 +52,4 @@ class NewsManager(ResourceManager):
         if response is True:
             response = {self.container: self.redmine.news.filter(**self.params)[0].raw()}
 
-        return super(NewsManager, self)._process_create_response(request, response)
+        return super()._process_create_response(request, response)
