@@ -13,11 +13,11 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_supports_url_retrieval(self):
         self.response.json.return_value = responses['project']['get']
-        self.assertEqual(self.redmine.project.get(1).url, '{0}/projects/foo'.format(self.url))
+        self.assertEqual(self.redmine.project.get(1).url, f'{self.url}/projects/foo')
 
     def test_supports_export_url_retrieval(self):
         self.response.json.return_value = responses['issue']['get']
-        self.assertEqual(self.redmine.issue.get(1).export_url('pdf'), '{0}/issues/1.pdf'.format(self.url))
+        self.assertEqual(self.redmine.issue.get(1).export_url('pdf'), f'{self.url}/issues/1.pdf')
         self.response.json.return_value = responses['attachment']['get']
         self.assertEqual(self.redmine.attachment.get(1).export_url('pdf'), None)
 
@@ -276,7 +276,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_project_url(self):
         self.response.json.return_value = responses['project']['get']
-        self.assertEqual(self.redmine.project.get(1).url, '{0}/projects/foo'.format(self.url))
+        self.assertEqual(self.redmine.project.get(1).url, f'{self.url}/projects/foo')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_project_export(self):
@@ -448,7 +448,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_issue_url(self):
         self.response.json.return_value = responses['issue']['get']
-        self.assertEqual(self.redmine.issue.get(1).url, '{0}/issues/1'.format(self.url))
+        self.assertEqual(self.redmine.issue.get(1).url, f'{self.url}/issues/1')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_issue_export(self):
@@ -534,7 +534,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_time_entry_url(self):
         self.response.json.return_value = responses['time_entry']['get']
-        self.assertEqual(self.redmine.time_entry.get(1).url, '{0}/time_entries/1'.format(self.url))
+        self.assertEqual(self.redmine.time_entry.get(1).url, f'{self.url}/time_entries/1')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_time_entry_export(self):
@@ -563,7 +563,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
         self.response.json.return_value = responses['enumeration']['filter']
         self.assertEqual(
             self.redmine.enumeration.filter(resource='time_entry_activities')[0].url,
-            '{0}/enumerations/1/edit'.format(self.url)
+            f'{self.url}/enumerations/1/edit'
         )
 
     def test_attachment_version(self):
@@ -610,7 +610,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_attachment_url(self):
         self.response.json.return_value = responses['attachment']['get']
-        self.assertEqual(self.redmine.attachment.get(1).url, '{0}/attachments/1'.format(self.url))
+        self.assertEqual(self.redmine.attachment.get(1).url, f'{self.url}/attachments/1')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_attachment_download(self):
@@ -684,7 +684,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_file_url(self):
         self.response.json.return_value = responses['attachment']['get']
-        self.assertEqual(self.redmine.file.get(1).url, '{0}/attachments/1'.format(self.url))
+        self.assertEqual(self.redmine.file.get(1).url, f'{self.url}/attachments/1')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_file_download(self):
@@ -706,10 +706,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
         """Test getting a wiki page with special char in title."""
         self.response.json.return_value = responses['wiki_page']['get_special']
         wiki_page = self.redmine.wiki_page.get('Foo%Bar', project_id=1)
-        self.assertEqual(
-            self.patch_requests.call_args[0][1],
-            '{0}/projects/1/wiki/Foo%25Bar.json'.format(self.url)
-        )
+        self.assertEqual(self.patch_requests.call_args[0][1], f'{self.url}/projects/1/wiki/Foo%25Bar.json')
         self.assertEqual(wiki_page.title, 'Foo%Bar')
         self.assertEqual(wiki_page.url, 'http://foo.bar/projects/1/wiki/Foo%25Bar')
 
@@ -730,10 +727,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
         self.response.status_code = 201
         self.response.json.return_value = responses['wiki_page']['get_special']
         wiki_page = self.redmine.wiki_page.create(project_id='foo', title='Foo%Bar')
-        self.assertEqual(
-            self.patch_requests.call_args[0][1],
-            '{0}/projects/foo/wiki/Foo%25Bar.json'.format(self.url)
-        )
+        self.assertEqual(self.patch_requests.call_args[0][1], f'{self.url}/projects/foo/wiki/Foo%25Bar.json')
         self.assertEqual(wiki_page.title, 'Foo%Bar')
 
     def test_wiki_page_delete(self):
@@ -797,10 +791,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_wiki_page_url(self):
         self.response.json.return_value = responses['wiki_page']['get']
-        self.assertEqual(
-            self.redmine.wiki_page.get('Foo', project_id='Foo').url,
-            '{0}/projects/Foo/wiki/Foo'.format(self.url)
-        )
+        self.assertEqual(self.redmine.wiki_page.get('Foo', project_id='Foo').url, f'{self.url}/projects/Foo/wiki/Foo')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_wiki_page_export(self):
@@ -865,7 +856,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_project_membership_url(self):
         self.response.json.return_value = responses['project_membership']['get']
-        self.assertEqual(self.redmine.project_membership.get(1).url, '{0}/memberships/1'.format(self.url))
+        self.assertEqual(self.redmine.project_membership.get(1).url, f'{self.url}/memberships/1')
 
     def test_issue_category_version(self):
         self.assertEqual(self.redmine.issue_category.resource_class.redmine_version, '1.3')
@@ -911,7 +902,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_issue_category_url(self):
         self.response.json.return_value = responses['issue_category']['get']
-        self.assertEqual(self.redmine.issue_category.get(1).url, '{0}/issue_categories/1'.format(self.url))
+        self.assertEqual(self.redmine.issue_category.get(1).url, f'{self.url}/issue_categories/1')
 
     def test_issue_relation_version(self):
         self.assertEqual(self.redmine.issue_relation.resource_class.redmine_version, '1.3')
@@ -956,7 +947,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_issue_relation_url(self):
         self.response.json.return_value = responses['issue_relation']['get']
-        self.assertEqual(self.redmine.issue_relation.get(1).url, '{0}/relations/1'.format(self.url))
+        self.assertEqual(self.redmine.issue_relation.get(1).url, f'{self.url}/relations/1')
 
     def test_version_version(self):
         self.assertEqual(self.redmine.version.resource_class.redmine_version, '1.3')
@@ -1007,7 +998,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_version_url(self):
         self.response.json.return_value = responses['version']['get']
-        self.assertEqual(self.redmine.version.get(1).url, '{0}/versions/1'.format(self.url))
+        self.assertEqual(self.redmine.version.get(1).url, f'{self.url}/versions/1')
 
     def test_user_version(self):
         self.assertEqual(self.redmine.user.resource_class.redmine_version, '1.1')
@@ -1116,7 +1107,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_user_url(self):
         self.response.json.return_value = responses['user']['get']
-        self.assertEqual(self.redmine.user.get(1).url, '{0}/users/1'.format(self.url))
+        self.assertEqual(self.redmine.user.get(1).url, f'{self.url}/users/1')
 
     def test_group_version(self):
         self.assertEqual(self.redmine.group.resource_class.redmine_version, '2.1')
@@ -1185,7 +1176,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_group_url(self):
         self.response.json.return_value = responses['group']['get']
-        self.assertEqual(self.redmine.group.get(1).url, '{0}/groups/1'.format(self.url))
+        self.assertEqual(self.redmine.group.get(1).url, f'{self.url}/groups/1')
 
     def test_role_version(self):
         self.assertEqual(self.redmine.role.resource_class.redmine_version, '1.4')
@@ -1206,7 +1197,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_role_url(self):
         self.response.json.return_value = responses['role']['get']
-        self.assertEqual(self.redmine.role.get(1).url, '{0}/roles/1'.format(self.url))
+        self.assertEqual(self.redmine.role.get(1).url, f'{self.url}/roles/1')
 
     def test_news_version(self):
         self.assertEqual(self.redmine.news.resource_class.redmine_version, '1.1')
@@ -1262,7 +1253,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_news_url(self):
         self.response.json.return_value = responses['news']['filter']
-        self.assertEqual(self.redmine.news.filter(project_id=1)[0].url, '{0}/news/2'.format(self.url))
+        self.assertEqual(self.redmine.news.filter(project_id=1)[0].url, f'{self.url}/news/2')
 
     @mock.patch('redminelib.open', mock.mock_open(), create=True)
     def test_news_export(self):
@@ -1297,7 +1288,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_issue_status_url(self):
         self.response.json.return_value = responses['issue_status']['all']
-        self.assertEqual(self.redmine.issue_status.all()[0].url, '{0}/issue_statuses/1/edit'.format(self.url))
+        self.assertEqual(self.redmine.issue_status.all()[0].url, f'{self.url}/issue_statuses/1/edit')
 
     def test_tracker_version(self):
         self.assertEqual(self.redmine.tracker.resource_class.redmine_version, '1.3')
@@ -1318,7 +1309,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_tracker_url(self):
         self.response.json.return_value = responses['tracker']['all']
-        self.assertEqual(self.redmine.tracker.all()[0].url, '{0}/trackers/1/edit'.format(self.url))
+        self.assertEqual(self.redmine.tracker.all()[0].url, f'{self.url}/trackers/1/edit')
 
     def test_query_version(self):
         self.assertEqual(self.redmine.query.resource_class.redmine_version, '1.3')
@@ -1339,7 +1330,7 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_query_url(self):
         self.response.json.return_value = responses['query']['all']
-        self.assertEqual(self.redmine.query.all()[0].url, '{0}/projects/0/issues?query_id=1'.format(self.url))
+        self.assertEqual(self.redmine.query.all()[0].url, f'{self.url}/projects/0/issues?query_id=1')
 
     def test_custom_field_version(self):
         self.assertEqual(self.redmine.custom_field.resource_class.redmine_version, '2.4')
@@ -1374,4 +1365,4 @@ class StandardResourcesTestCase(BaseRedmineTestCase):
 
     def test_custom_field_url(self):
         self.response.json.return_value = responses['custom_field']['all']
-        self.assertEqual(self.redmine.custom_field.all()[0].url, '{0}/custom_fields/1/edit'.format(self.url))
+        self.assertEqual(self.redmine.custom_field.all()[0].url, f'{self.url}/custom_fields/1/edit')
